chore: added documentation
